Convert the repo to a monorepo and split the WebSocket event streaming to own package (#123)
Converts the repo to a monorepo and splits the WebSocket event streaming functionality (MailpitEvents) out of mailpit-api into a new, standalone package mailpit-ws.

- Restructured as a monorepo under mailpit-api and mailpit-ws
- mailpit-ws is a dedicated TypeScript WebSocket client for Mailpit real-time events
- Updated tsconfig, eslint, and vitest configs to support the monorepo layout
- Added fetchOptions support in the mailpit-api constructor
- Improved baseURL normalization (trailing slash handling, malformed URLs)
