+++ .github/dependabot.yml
@@ -17,3 +17,20 @@ updates:
       github-actions:
         patterns:
           - "*"
+  - package-ecosystem: "npm"
+    directory: "/"
+    schedule:
+      interval: "monthly"
+      day: "friday"
+      time: "21:00"
+    open-pull-requests-limit: 5
+    assignees:
+      - "mpspahr"
+    labels:
+      - "dependencies"
+    commit-message:
+      prefix: "chore(deps):"
+    groups:
+      npm-dependencies:
+        patterns:
+          - "*"
