+++ .github/workflows/npm-publish.yml
@@ -1,10 +1,12 @@
-# This workflow is triggered on the creation of a new release.
-# It runs tests and, if successful, builds and publishes the package to NPM.
+# This workflow is triggered when a GitHub release is published.
+# It runs tests and, if successful, builds and publishes the tagged package to NPM.
+# Tags must follow the format: mailpit-api@x.y.z or mailpit-ws@x.y.z
+# Pre-releases are published to npm with the "next" dist-tag instead of "latest".
 name: Publish Release to NPM
 
 on:
   release:
-    types: [created]
+    types: [published]
 
 concurrency:
   group: "npm-publish-${{ github.ref }}"
@@ -34,4 +36,28 @@ jobs:
           registry-url: https://registry.npmjs.org/
       - run: npm ci
       - run: npm run build
-      - run: npm publish --tag legacy
+      - name: Determine package to publish
+        id: pkg
+        run: |
+          TAG="${{ github.event.release.tag_name }}"
+          PACKAGE_NAME="${TAG%@*}"
+          VERSION="${TAG##*@}"
+          echo "tag=$TAG" >> "$GITHUB_OUTPUT"
+          echo "name=$PACKAGE_NAME" >> "$GITHUB_OUTPUT"
+          echo "version=$VERSION" >> "$GITHUB_OUTPUT"
+          echo "Publishing $PACKAGE_NAME@$VERSION"
+      - name: Publish to npm
+        run: |
+          if [ "${{ github.event.release.prerelease }}" = "true" ]; then
+            npm publish --workspace packages/${{ steps.pkg.outputs.name }} --provenance --access public --tag next
+          else
+            npm publish --workspace packages/${{ steps.pkg.outputs.name }} --provenance --access public
+          fi
+
+  call-docs:
+    needs: publish
+    permissions:
+      contents: read
+      pages: write
+      id-token: write
+    uses: ./.github/workflows/deploy-docs.yml
