+++ .gitignore
@@ -90,6 +90,7 @@ out
 # Nuxt.js build / generate output
 .nuxt
 dist
+packages/*/dist
 
 # Gatsby files
 .cache/
